requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
python-dotenv==1.0.1
redis==5.0.7
heyoo==0.1.7
gspread==6.1.2
oauth2client==4.1.3
Pillow==10.4.0
gunicorn==23.0.0
requests==2.32.3